+++ libs/cgse-core/pyproject.toml
@@ -44,6 +44,12 @@ filterwarnings = [
     "ignore::DeprecationWarning"
 ]
 
+[tool.coverage.run]
+omit = [
+    "tests/*",
+    "conftest.py",
+]
+
 [tool.hatch.build.targets.sdist]
 exclude = [
     "/tests",
