+++ libs/cgse-core/tests/test_confman_setups.py
@@ -4,7 +4,7 @@ from egse.env import set_conf_repo_location
 from egse.setup import load_setup
 
 
-def test_private_caching_functions():
+def test_private_caching_functions(default_env):
 
     from egse.confman import _get_cached_setup_info
     from egse.confman import _populate_cached_setup_info
