+++ libs/cgse-core/tests/test_settings_core.py
