+++ pyproject.toml
@@ -14,6 +14,23 @@ dependencies = [
     "tomlkit>=0.13.2",
 ]
 
+[tool.pytest.ini_options]
+pythonpath = "src"
+testpaths = [
+    "libs/cgse-common/tests",
+    "libs/cgse-core/tests",
+]
+addopts = "-ra --cov --cov-report html"
+filterwarnings = [
+    "ignore::DeprecationWarning"
+]
+
+[tool.coverage.run]
+omit = [
+    "*/tests/*",
+    "conftest.py",
+]
+
 [tool.uv.sources]
 cgse-common = { workspace = true }
 cgse-core = { workspace = true }
